Update transform/models/intermediate/clearinghouse/int_clearinghouse__five_minute_station_agg.sql
Thanks for the thorough breakdown @ian-r-rose! My only concern using nullifzero with the occupancy is that the speed result is null when I wanted to return 0. Since these values are using the simplified speed formula I think it would be ok to return null with the knowledge that null speed values correspond to the following situations:

- No flow value
- No occupancy value
- Occupancy value of 0

Co-authored-by: Ian Rose <ian.rose@innovation.ca.gov>